12.0.2
- Fixed issue that could cause fromTo() tweens in a TimelineLite/Max
not to revert to their initial values (even before the "from" part)
when the timeline is rewound past the startTime of the fromTo() tween.

- Fixed issue that could cause the onFail event of a child of a
LoaderMax not to propagate up through the LoaderMax instance.

- Added data to the HTTP_STATUS LoaderEvent so that you can check the
status value directly on the LoaderEvent.text and the original event is
referenced on the LoaderEvent.data.

- XMLLoader now recognizes the allowMalformedURL special property.