12.0.11
- The virtual playhead of a TimelineLite and TimelineMax instance now
adjusts backwards if necessary when children are removed. For example,
if the timeline has completely played and then all children are
removed, its virtual playhead will return to 0.

- Fixed issue that could cause all callbacks to be removed from a
TimelineMax if null was passed to its removeCallback() method as the
first parameter.

- Fixed issue in TimelineMax that could cause a problem if its pause()
method is called from within its onRepeat.

- Fixed an issue with filter tweens that could cause an error if
multiple filters are tweened on an object and the index is defined for
each and there is no filters at those indexes yet.